+++ CMakeLists.txt
@@ -626,6 +626,7 @@ set(SRC_SHP)
 set(SRC_GUI)
 set(SRC_UTL)
 set(SRC_SKETCH)
+set(SRC_SCR)
 set(SRC_OTHER)
 foreach(f ${srcs})
   get_filename_component(_src_name ${f} NAME)
@@ -633,10 +634,12 @@ foreach(f ${srcs})
     list(APPEND SRC_SHP ${f})
   elseif(_src_name MATCHES "^gui_")
     list(APPEND SRC_GUI ${f})
-  elseif(_src_name MATCHES "^utl_")
+  elseif(_src_name MATCHES "^utl(_|\\.)")
     list(APPEND SRC_UTL ${f})
   elseif(_src_name MATCHES "^sketch_")
     list(APPEND SRC_SKETCH ${f})
+  elseif(_src_name MATCHES "^scr_")
+    list(APPEND SRC_SCR ${f})
   else()
     list(APPEND SRC_OTHER ${f})
   endif()
@@ -645,6 +648,7 @@ source_group("src\\shp" FILES ${SRC_SHP})
 source_group("src\\gui" FILES ${SRC_GUI})
 source_group("src\\utl" FILES ${SRC_UTL})
 source_group("src\\sketch" FILES ${SRC_SKETCH})
+source_group("src\\scr" FILES ${SRC_SCR})
 source_group("src" FILES ${SRC_OTHER})
 source_group("web" FILES ${WEB_FILES_REL})
