+++ src/gui.cpp
@@ -25,9 +25,9 @@
 #include "utl_geom.h"
 #include "imgui.h"
 #include "utl_log.h"
-#include "lua_console.h"
+#include "scr_lua_console.h"
 #include "occt_view.h"
-#include "python_console.h"
+#include "scr_python_console.h"
 #include "shp_info.h"
 #include "sketch.h"
 #include "utl.h"
