+++ src/occt_view.cpp
@@ -53,7 +53,7 @@
 #include "shp_create.h"
 #include "sketch.h"
 #include "sketch_json.h"
-#include "types.h"
+#include "utl_types.h"
 #include "utl.h"
 #include "utl_json.h"
 #include "utl_occt.h"
