+++ src/utl_geom.h
@@ -13,7 +13,7 @@
 #include <vector>
 
 #include "utl_dbg.h"
-#include "types.h"
+#include "utl_types.h"
 
 class gp_Pln;
 class gp_Pnt2d;
