+++ src/utl_types.h
@@ -64,4 +64,4 @@ enum class Export_format
   Ply
 };
 
-#include "types.inl"
+#include "utl_types.inl"
