+++ src/utl_types.inl
